Update README.md
This project applies a Random Forest Regression model to a dataset of the top Spotify songs of 2023. The goal is to predict certain musical attributes and trends using the available data.Overview
Dataset
This dataset contains a comprehensive list of the most famous songs of 2023, as listed on Spotify. It offers a wealth of features beyond what is typically available in similar datasets, providing insights into each song's attributes, popularity, and presence on various music platforms.